Re-arranged all components
