+++ reactfe/src/App.js
@@ -1,23 +1,11 @@
 import React, { Component } from 'react'
 import './App.css';
-// import {BrowserRouter, Switch, Route} from 'react-router-dom';
-import {Provider} from './Components/provider'
-import {Container} from 'reactstrap'
-import HeaderComponent from './Components/HeaderComponent'
-import BodyComponent from './Components/BodyComponent'
+import RedditApp from './RedditApp'
 
-
-class App extends Component {
-  render() {
-    return (
-      <Provider>
-          <Container>
-            <HeaderComponent />
-            <BodyComponent />
-          </Container>
-      </Provider>
-    );
-  }
+export default class App extends Component {
+	render() {
+		return (
+			<RedditApp />
+		);
+	}
 }
-
-export default App;
