--- reactfe/src/Components/HeaderComponent/NavbarComponent/index.js
@@ -1,46 +1,0 @@
-import React from 'react';
-import Context from '../../provider'
-import {
-    Collapse,
-    Navbar, 
-    NavbarToggler, 
-    NavbarBrand, 
-    Nav, 
-    NavItem, 
-    NavLink,
-} from 'reactstrap';
-import LoginComponent from './LoginComponent'
-import SignupComponent from './SignupComponent'
-import SignoutComponent from './SignoutComponent'
-
-export default class NavbarComponent extends React.Component {
-    render() {
-        return (
-            <Context.Consumer>
-                {context => {
-                    return (
-                        <Navbar color="light" light expand="md">
-                            <NavbarBrand href="/">Reddit</NavbarBrand>
-                                <NavbarToggler onClick={() => context.toggleNavbar()} />
-                                <Collapse isOpen={context.navbarOpen} navbar>
-                                <Nav className="ml-auto" navbar>
-                                    {context.loggedIn ? 
-                                        <React.Fragment>
-                                            <NavItem><NavLink href={'/u/' + context.username.toLowerCase() + '/'}>{context.username}</NavLink></NavItem>
-                                            <NavItem><SignoutComponent /></NavItem>
-                                        </React.Fragment>
-                                        :
-                                        <React.Fragment>
-                                            <NavItem><LoginComponent /></NavItem>
-                                            <NavItem><SignupComponent /></NavItem>
-                                        </React.Fragment>
-                                    }
-                                </Nav>
-                            </Collapse>
-                        </Navbar>
-                    )   
-                }}
-            </Context.Consumer>
-        )
-    }
-}
