+++ reactfe/src/RedditApp/BodyComponent/ContentComponent/Profile/ProfileComments/index.js
@@ -1,0 +1,52 @@
+import React, {Component} from 'react'
+import {
+    Card,
+    CardHeader,
+    CardBody,
+    CardText,
+    CardLink,
+    Row,
+    Col,
+} from 'reactstrap'
+import '../../assets/CommentTemplate'
+import CommentTemplate from '../../assets/CommentTemplate';
+
+export default class ProfileComments extends Component {
+    constructor(props) {
+        super(props)
+        this.state = {comments: []}
+    }
+
+    componentDidMount() {
+        fetch(`/api/reddit/u/${this.props.username}/comments/`)
+        .then(data => data.json())
+        .then(json => {
+            this.setState({
+                comments: json,
+            })
+        })
+    }
+
+    render() {
+        return (
+            <React.Fragment>
+            {this.state.comments.map((comment) => {
+                    return (
+                        <Row key={comment.id}>
+                            <Col>
+                                <CommentTemplate
+                                    content={comment.content}                            
+                                    votes={comment.votes}
+                                    postid={comment.parent_post.id}
+                                    subreddit={comment.parent_post.subreddit.name}
+                                    subredditlink={true}
+                                />
+                            </Col>
+                        </Row>
+                    )
+                }
+            )}
+            </React.Fragment>
+        )
+    }
+}
