+++ reactfe/src/RedditApp/BodyComponent/ContentComponent/assets/NewPost/index.js
