+++ reactfe/src/RedditApp/BodyComponent/InfoComponent/InfoCardComponent/index.js
