+++ reactfe/src/RedditApp/BodyComponent/InfoComponent/index.js
