+++ reactfe/src/RedditApp/BodyComponent/index.js
