+++ reactfe/src/RedditApp/HeaderComponent/NavbarComponent/Logout/index.js
@@ -1,0 +1,19 @@
+import React, {Component} from 'react'
+import Context from '../../../../provider'
+import {
+    Button,
+} from 'reactstrap'
+
+export default class Logout extends Component {
+    render() {
+        return (
+            <Context.Consumer>
+                {context => {
+                    return (
+                        <Button color='alert' onClick={() => context.toggleLoggedIn()}>Signout</Button>
+                    )
+                }}
+            </Context.Consumer>
+        )
+    }
+}
