+++ reactfe/src/RedditApp/HeaderComponent/index.js
