+++ webpack-stats.dev.json
@@ -1,1 +1,1 @@
-{"status":"done","publicPath":"http://localhost:3000/","error":"ModuleError","message":"\n  Line 74:  Parsing error: Unexpected token, expected ,\n\n  72 |         super(props)(\n  73 |         this.state = {temp:false}\n> 74 |     }\n     |     ^\n  75 | \n  76 |     componentWillReceiveProps() {\n  77 |         this.setState(prev => ({\n\n","chunks":{"main":[{"name":"static/js/bundle.js","publicPath":"http://localhost:3000/static/js/bundle.js","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\static\\js\\bundle.js"},{"name":"0.b8a37f07a7a94c989f2d.hot-update.js","publicPath":"http://localhost:3000/0.b8a37f07a7a94c989f2d.hot-update.js","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\0.b8a37f07a7a94c989f2d.hot-update.js"},{"name":"static/js/bundle.js.map","publicPath":"http://localhost:3000/static/js/bundle.js.map","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\static\\js\\bundle.js.map"},{"name":"0.b8a37f07a7a94c989f2d.hot-update.js.map","publicPath":"http://localhost:3000/0.b8a37f07a7a94c989f2d.hot-update.js.map","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\0.b8a37f07a7a94c989f2d.hot-update.js.map"}]}}
+{"status":"error","publicPath":"http://localhost:3000/","error":"ModuleBuildError","message":"Module build failed: Error: ENOENT: no such file or directory, open 'C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\src\\Components\\provider.js'","chunks":{"main":[{"name":"static/js/bundle.js","publicPath":"http://localhost:3000/static/js/bundle.js","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\static\\js\\bundle.js"},{"name":"0.b8a37f07a7a94c989f2d.hot-update.js","publicPath":"http://localhost:3000/0.b8a37f07a7a94c989f2d.hot-update.js","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\0.b8a37f07a7a94c989f2d.hot-update.js"},{"name":"static/js/bundle.js.map","publicPath":"http://localhost:3000/static/js/bundle.js.map","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\static\\js\\bundle.js.map"},{"name":"0.b8a37f07a7a94c989f2d.hot-update.js.map","publicPath":"http://localhost:3000/0.b8a37f07a7a94c989f2d.hot-update.js.map","path":"C:\\Users\\Sreekar Mouli\\Documents\\Mission RnD Summer\\New folder\\reddit\\reactfe\\0.b8a37f07a7a94c989f2d.hot-update.js.map"}]}}
